callautomation-live-transcription/requirements.txt
quart==0.19.6
azure-communication-callautomation==1.3.0b1
azure-eventgrid==4.11.0
azure-core==1.28.0
requests==2.31.0
websockets==11.0.3